+++ examples/nextjs/pages/index.tsx
@@ -15,9 +15,9 @@ import {
   orderedList as elementOrderedList,
   todoList as elementTodoList,
   todoListItem as elementTodoListItem,
-  LiRenderer,
-  UlRenderer,
-  OlRenderer,
+  ListItemRenderer,
+  ListRenderer,
+  OrderedListRenderer,
   TodoListRenderer,
   TodoListItemRenderer,
 } from '@bento-editor/element-list';
@@ -312,9 +312,9 @@ const Home: NextPage = () => {
     () => ({
       paragraph: ParagraphRenderer,
       heading: HeadingRenderer,
-      list: UlRenderer,
-      'ordered-list': OlRenderer,
-      'list-item': LiRenderer,
+      list: ListRenderer,
+      'ordered-list': OrderedListRenderer,
+      'list-item': ListItemRenderer,
       'todo-list': TodoListRenderer,
       'todo-list-item': TodoListItemRenderer,
       format: TextFormatRenderer,
