+++ packages/element-list/src/item/index.ts
@@ -2,6 +2,7 @@ import { Element } from '@bento-editor/core';
 import attributes, { Attributes } from './attributes';
 import editable from './editable';
 import toolbox from './toolbox';
+export * from './renderer';
 
 const element: Element<Attributes> = {
   type: 'list-item',
