+++ packages/element-list/src/item/renderer/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { RendererProps } from '@bento-editor/core';
 
-export const LiRenderer: React.FC<RendererProps> = ({ children }) => {
+export const ListItemRenderer: React.FC<RendererProps> = ({ children }) => {
   return <li>{children}</li>;
 };
