+++ packages/element-list/src/list/renderer/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { RendererProps } from '@bento-editor/core';
-import { Attributes } from '../list/attributes';
+import { Attributes } from '../attributes';
 
-export const UlRenderer: React.FC<RendererProps<Attributes>> = ({
+export const ListRenderer: React.FC<RendererProps<Attributes>> = ({
   children,
   attributes,
 }) => {
