+++ packages/element-list/src/ordered/index.ts
@@ -2,6 +2,7 @@ import { Element, helpers } from '@bento-editor/core';
 import attributes, { Attributes } from './attributes';
 import editable from './editable';
 import toolbox from './toolbox';
+export * from './renderer';
 
 const element: Element<Attributes> = {
   type: 'ordered-list',
