--- packages/element-list/src/renderer/index.ts
@@ -1,5 +1,0 @@
-export { UlRenderer } from './UlRenderer';
-export { OlRenderer } from './OlRenderer';
-export { LiRenderer } from './LiRenderer';
-export { TodoListRenderer } from './TodoListRenderer';
-export { TodoListItemRenderer } from './TodoListItemRenderer';
